+++ tests/ErrorHandlingTest.php
@@ -9,12 +9,14 @@ use BlueMvc\Core\Tests\Helpers\TestApplications\BasicTestApplication;
 use BlueMvc\Core\Tests\Helpers\TestControllers\ActionResultTestController;
 use BlueMvc\Core\Tests\Helpers\TestControllers\BasicTestController;
 use BlueMvc\Core\Tests\Helpers\TestControllers\ErrorTestController;
+use BlueMvc\Core\Tests\Helpers\TestControllers\ErrorTraitTestController;
 use BlueMvc\Core\Tests\Helpers\TestControllers\ExceptionTestController;
 use BlueMvc\Core\Tests\Helpers\TestRequests\BasicTestRequest;
 use BlueMvc\Core\Tests\Helpers\TestResponses\BasicTestResponse;
 use BlueMvc\Core\Tests\Helpers\TestViewRenderers\BasicTestViewRenderer;
 use DataTypes\FilePath;
 use DataTypes\Url;
+use DataTypes\UrlPath;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -23,132 +25,72 @@ use PHPUnit\Framework\TestCase;
 class ErrorHandlingTest extends TestCase
 {
     /**
-     * Test server error in debug mode.
+     * Test error handling.
+     *
+     * @dataProvider errorHandlingDataProvider
+     *
+     * @param bool        $isDebug              If true, application is in debug mode. False otherwise.
+     * @param string      $path                 The path.
+     * @param string|null $errorControllerClass The error controller class or null if no error controller.
+     * @param array       $expectedHeaders      The expected headers.
+     * @param string      $expectedContent      The expected content.
+     * @param int         $expectedStatusCode   The expected status code.
      */
-    public function testServerErrorInDebugMode()
+    public function testErrorHandling($isDebug, $path, $errorControllerClass, array $expectedHeaders, $expectedContent, $expectedStatusCode)
     {
-        $this->application->setDebug(true);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/exception/'), new Method('GET'));
+        if ($errorControllerClass !== null) {
+            $this->application->setErrorControllerClass($errorControllerClass);
+        }
+        $this->application->setDebug($isDebug);
+        $request = new BasicTestRequest(Url::parse('https://www.example.com/')->withPath(UrlPath::parse($path)), new Method('GET'));
         $response = new BasicTestResponse();
         $this->application->run($request, $response);
 
-        self::assertContains('<title>Exception was thrown.</title>', $response->getContent());
-        self::assertSame(StatusCode::INTERNAL_SERVER_ERROR, $response->getStatusCode()->getCode());
+        if ($expectedContent === '') {
+            self::assertSame('', $response->getContent());
+        } else {
+            self::assertContains($expectedContent, $response->getContent());
+        }
+
+        self::assertSame($expectedHeaders, iterator_to_array($response->getHeaders()));
+        self::assertSame($expectedStatusCode, $response->getStatusCode()->getCode());
     }
 
     /**
-     * Test server error in release mode.
+     * Data provider for error handling tests.
+     *
+     * @return array The data.
      */
-    public function testServerErrorInReleaseMode()
+    public function errorHandlingDataProvider()
     {
-        $this->application->setDebug(false);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/exception/'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertSame('', $response->getContent());
-        self::assertSame(StatusCode::INTERNAL_SERVER_ERROR, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test server error with error controller set.
-     */
-    public function testServerErrorWithErrorControllerSet()
-    {
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/exception/'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertSame('<html><body><h1>Request Failed: Error: 500, Exception: LogicException, ExceptionMessage: Exception was thrown.</h1></body></html>', $response->getContent());
-        self::assertSame(StatusCode::INTERNAL_SERVER_ERROR, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test server error with error controller returning null.
-     */
-    public function testServerErrorWithErrorControllerReturningNull()
-    {
-        $this->application->setDebug(true);
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/exception/domainException'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertContains('<title>DomainException was thrown.</title>', $response->getContent());
-        self::assertSame(StatusCode::INTERNAL_SERVER_ERROR, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test not found error with error controller set.
-     */
-    public function testNotFoundErrorWithErrorControllerSet()
-    {
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/non-existing'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertSame('<html><body><h1>Request Failed: Error: 404</h1></body></html>', $response->getContent());
-        self::assertSame(StatusCode::NOT_FOUND, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test not found error with error controller set 2.
-     */
-    public function testNotFoundErrorWithErrorControllerSet2()
-    {
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/actionresult/notfound'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertSame('<html><body><h1>Request Failed: Error: 404</h1></body></html>', $response->getContent());
-        self::assertSame(StatusCode::NOT_FOUND, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test normal page with error controller set.
-     */
-    public function testNormalPageWithErrorControllerSet()
-    {
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertSame('Hello World!', $response->getContent());
-        self::assertSame(StatusCode::OK, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test error controller throwing exception in non-debug mode.
-     */
-    public function testErrorControllerThrowingExceptionInNonDebugMode()
-    {
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/actionresult/forbidden'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertSame('', $response->getContent());
-        self::assertSame(StatusCode::INTERNAL_SERVER_ERROR, $response->getStatusCode()->getCode());
-    }
-
-    /**
-     * Test error controller throwing exception in debug mode.
-     */
-    public function testErrorControllerThrowingExceptionInDebugMode()
-    {
-        $this->application->setDebug(true);
-        $this->application->setErrorControllerClass(ErrorTestController::class);
-        $request = new BasicTestRequest(Url::parse('https://www.example.com/actionresult/forbidden'), new Method('GET'));
-        $response = new BasicTestResponse();
-        $this->application->run($request, $response);
-
-        self::assertContains('RuntimeException', $response->getContent());
-        self::assertContains('Exception thrown from 403 action.', $response->getContent());
-        self::assertSame(StatusCode::INTERNAL_SERVER_ERROR, $response->getStatusCode()->getCode());
+        return [
+            [true, '/exception/', null, [], '<title>Exception was thrown.</title>', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/exception/', null, [], '', StatusCode::INTERNAL_SERVER_ERROR],
+            [true, '/exception/', ErrorTestController::class, [], '<html><body><h1>Request Failed: Error: 500, Exception: LogicException, ExceptionMessage: Exception was thrown.</h1></body></html>', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/exception/', ErrorTestController::class, [], '<html><body><h1>Request Failed: Error: 500, Exception: LogicException, ExceptionMessage: Exception was thrown.</h1></body></html>', StatusCode::INTERNAL_SERVER_ERROR],
+            [true, '/exception/domainException', ErrorTestController::class, [], '<title>DomainException was thrown.</title>', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/exception/domainException', ErrorTestController::class, [], '', StatusCode::INTERNAL_SERVER_ERROR],
+            [true, '/exception/non-existing', ErrorTestController::class, ['X-Error-PreActionEvent' => '1', 'X-Error-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [false, '/exception/non-existing', ErrorTestController::class, ['X-Error-PreActionEvent' => '1', 'X-Error-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [true, '/exception/actionresult/notfound', ErrorTestController::class, ['X-Error-PreActionEvent' => '1', 'X-Error-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [false, '/exception/actionresult/notfound', ErrorTestController::class, ['X-Error-PreActionEvent' => '1', 'X-Error-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [true, '/', ErrorTestController::class, [], 'Hello World!', StatusCode::OK],
+            [false, '/', ErrorTestController::class, [], 'Hello World!', StatusCode::OK],
+            [true, '/actionresult/forbidden', ErrorTestController::class, [], 'Exception thrown from 403 action.', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/actionresult/forbidden', ErrorTestController::class, [], '', StatusCode::INTERNAL_SERVER_ERROR],
+            [true, '/exception/', ErrorTraitTestController::class, [], '<html><body><h1>Request Failed: Error: 500, Exception: LogicException, ExceptionMessage: Exception was thrown.</h1></body></html>', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/exception/', ErrorTraitTestController::class, [], '<html><body><h1>Request Failed: Error: 500, Exception: LogicException, ExceptionMessage: Exception was thrown.</h1></body></html>', StatusCode::INTERNAL_SERVER_ERROR],
+            [true, '/exception/domainException', ErrorTraitTestController::class, [], '<title>DomainException was thrown.</title>', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/exception/domainException', ErrorTraitTestController::class, [], '', StatusCode::INTERNAL_SERVER_ERROR],
+            [true, '/exception/non-existing', ErrorTraitTestController::class, ['X-ErrorTrait-PreActionEvent' => '1', 'X-ErrorTrait-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [false, '/exception/non-existing', ErrorTraitTestController::class, ['X-ErrorTrait-PreActionEvent' => '1', 'X-ErrorTrait-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [true, '/exception/actionresult/notfound', ErrorTraitTestController::class, ['X-ErrorTrait-PreActionEvent' => '1', 'X-ErrorTrait-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [false, '/exception/actionresult/notfound', ErrorTraitTestController::class, ['X-ErrorTrait-PreActionEvent' => '1', 'X-ErrorTrait-PostActionEvent' => '1'], '<html><body><h1>Request Failed: Error: 404</h1></body></html>', StatusCode::NOT_FOUND],
+            [true, '/', ErrorTraitTestController::class, [], 'Hello World!', StatusCode::OK],
+            [false, '/', ErrorTraitTestController::class, [], 'Hello World!', StatusCode::OK],
+            [true, '/actionresult/forbidden', ErrorTraitTestController::class, [], 'Exception thrown from 403 action.', StatusCode::INTERNAL_SERVER_ERROR],
+            [false, '/actionresult/forbidden', ErrorTraitTestController::class, [], '', StatusCode::INTERNAL_SERVER_ERROR],
+        ];
     }
 
     /**