+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 - Added getClientIp method to Request.
+- Added ErrorControllerTrait.
+- Added isPreActionEventEnabled and isPostActionEventEnabled methods to Controller.
+- Prevented onPreActionEvent and onPreActionEvent methods from running on error controller if an exception occurred.
 
 ## 1.0.0 - 2018-02-03
 - First stable revision.